remove special case stralloc textcode functions write generic stralloc and array textcode wrapper functions change textcode API to use long instead of int add cescape fmt and scan functions to textcode add fmt_foldwhitespace to textcode

+++ textcode/fmt_html.c
@@ -3,7 +3,7 @@
 #include "str.h"
 #include "haveinline.h"
 
-unsigned int fmt_html(char* dest,const char* src,unsigned int len) {
+unsigned long fmt_html(char* dest,const char* src,unsigned long len) {
   register const unsigned char* s=(const unsigned char*) src;
   unsigned long written=0,i;
   const char* seq;
